refactor: Do not distinguish between Connection and ConnectionRaii
